Remove page component; inline navbar / footer rendering
